Lots of changes here (yes, I know that was a poor version control decision, but it's too late now).
Anyway, here's a breakdown:
- Ran Expo-Doctor to get all expo modules to correct version
- Navigation bar now uses icons instead of words (plus added color!)
- Calendar page now shows the complete scale in the correct order (it was previously starting with Relapse at the top, which was unintended). It also scrolls now.
- Most of the new work was done on my entry form. It now displays the options in columns (and auto-adjusts the number of columns based on screen size), it has headers for each level of the scale, and it scrolls!
- Oh, and CheckboxGroup was moved to its own components folder just so Expo would stop yelling at me about lack of an export.
- After putting in the full list of FASTER scale options, [date] was looking a bit long, so I moved the entire option list to its own file in the new data folder

Current known issue: The animations on the navigation bar are spilling down into the device's navigation bar. It's super obvious on my tablet, though I'm not sure how it looks on other devices. It's a minor thing, though, so I'm in no rush to fix it.

I've also run out of GitHub chat messages for the month. 6 days into the month. So yeah...